+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 - Scenario outline is reported with the line number from the `Scenario Outline` keyword (not the examples table anymore) ([77](https://github.com/ThoSe1990/cwt-cucumber/pull/77))
-- Undefined steps are reported as undefined and not skipped when a prior step fails ([77](https://github.com/ThoSe1990/cwt-cucumber/pull/77))
+- Undefined steps are always reported as undefined. Previously, if you used a skip tag, it wasn't reported at all ([77](https://github.com/ThoSe1990/cwt-cucumber/pull/77))
 
 ### Fixed
 - Scenario outline: Variable replacement in doc-strings and description ([77](https://github.com/ThoSe1990/cwt-cucumber/pull/77))